// frontend/src/openapi/models/EvalRunList.ts
export interface EvalRunList {
    /**
     *
     * @type {string}
     * @memberof EvalRunList
     */
    readonly id: string;
    /**
     *
     * @type {string}
     * @memberof EvalRunList
     */
    readonly project: string;
    /**
     *
     * @type {string}
     * @memberof EvalRunList
     */
    readonly name: string;
    /**
     *
     * @type {string}
     * @memberof EvalRunList
     */
    readonly description: string;
    /**
     *
     * @type {EvalRunDataSourceEnum}
     * @memberof EvalRunList
     */
    readonly dataSource: EvalRunDataSourceEnum;
    /**
     *
     * @type {string}
     * @memberof EvalRunList
     */
    readonly dataset: string | null;
    /**
     *
     * @type {string}
     * @memberof EvalRunList
     */
    readonly datasetName: string;
    /**
     *
     * @type {string}
     * @memberof EvalRunList
     */
    readonly capabilityName: string;
    /**
     *
     * @type {string}
     * @memberof EvalRunList
     */
    readonly capabilityId: string;
    /**
     *
     * @type {string}
     * @memberof EvalRunList
     */
    readonly origin: string;
    /**
     *
     * @type {string}
     * @memberof EvalRunList
     */
    readonly optimizerExperiment: string | null;
    /**
     *
     * @type {Status1b7Enum}
     * @memberof EvalRunList
     */
    readonly status: Status1b7Enum;
    /**
     *
     * @type {string}
     * @memberof EvalRunList
     */
    readonly error: string;
    /**
     *
     * @type {number}
     * @memberof EvalRunList
     */
    readonly maxItems: number;
    /**
     *
     * @type {number}
     * @memberof EvalRunList
     */
    readonly sampling: number;
    /**
     *
     * @type {number}
     * @memberof EvalRunList
     */
    readonly variantCount: number;
    /**
     *
     * @type {number}
     * @memberof EvalRunList
     */
    readonly evaluatorCount: number;
    /**
     *
     * @type {EvalRunProgress}
     * @memberof EvalRunList
     */
    readonly progress: EvalRunProgress | null;
    /**
     *
     * @type {Date}
     * @memberof EvalRunList
     */
    readonly createdAt: Date;
    /**
     *
     * @type {Date}
     * @memberof EvalRunList
     */
    readonly completedAt: Date | null;
}