// frontend/src/openapi/models/GenerateEvaluatorPromptRequestRequest.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Overbae API
 * Backend API for the Overclaw / Overmind platform.
 *
 * The version of the OpenAPI document: 2.0.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { ApplicableRoleEnum } from './ApplicableRoleEnum';
import {
    ApplicableRoleEnumFromJSON,
    ApplicableRoleEnumFromJSONTyped,
    ApplicableRoleEnumToJSON,
    ApplicableRoleEnumToJSONTyped,
} from './ApplicableRoleEnum';

/**
 *
 * @export
 * @interface GenerateEvaluatorPromptRequestRequest
 */
export interface GenerateEvaluatorPromptRequestRequest {
    /**
     *
     * @type {string}
     * @memberof GenerateEvaluatorPromptRequestRequest
     */
    description: string;
    /**
     *
     * @type {string}
     * @memberof GenerateEvaluatorPromptRequestRequest
     */
    capability?: string | null;
    /**
     *
     * @type {ApplicableRoleEnum}
     * @memberof GenerateEvaluatorPromptRequestRequest
     */
    applicableRole?: ApplicableRoleEnum;
}



/**
 * Check if a given object implements the GenerateEvaluatorPromptRequestRequest interface.
 */
export function instanceOfGenerateEvaluatorPromptRequestRequest(value: object): value is GenerateEvaluatorPromptRequestRequest {
    if (!('description' in value) || value['description'] === undefined) return false;
    return true;
}

export function GenerateEvaluatorPromptRequestRequestFromJSON(json: any): GenerateEvaluatorPromptRequestRequest {
    return GenerateEvaluatorPromptRequestRequestFromJSONTyped(json, false);
}

export function GenerateEvaluatorPromptRequestRequestFromJSONTyped(json: any, ignoreDiscriminator: boolean): GenerateEvaluatorPromptRequestRequest {
    if (json == null) {
        return json;
    }
    return {

        'description': json['description'],
        'capability': json['capability'] == null ? undefined : json['capability'],
        'applicableRole': json['applicable_role'] == null ? undefined : ApplicableRoleEnumFromJSON(json['applicable_role']),
    };
}

export function GenerateEvaluatorPromptRequestRequestToJSON(json: any): GenerateEvaluatorPromptRequestRequest {
    return GenerateEvaluatorPromptRequestRequestToJSONTyped(json, false);
}

export function GenerateEvaluatorPromptRequestRequestToJSONTyped(value?: GenerateEvaluatorPromptRequestRequest | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {

        'description': value['description'],
        'capability': value['capability'],
        'applicable_role': ApplicableRoleEnumToJSON(value['applicableRole']),
    };
}