// frontend/src/openapi/models/PaginatedEvalRunListList.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Overbae API
 * Backend API for the Overclaw / Overmind platform.
 *
 * The version of the OpenAPI document: 2.0.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { EvalRunList } from './EvalRunList';
import {
    EvalRunListFromJSON,
    EvalRunListFromJSONTyped,
    EvalRunListToJSON,
    EvalRunListToJSONTyped,
} from './EvalRunList';

/**
 *
 * @export
 * @interface PaginatedEvalRunListList
 */
export interface PaginatedEvalRunListList {
    /**
     *
     * @type {number}
     * @memberof PaginatedEvalRunListList
     */
    count: number;
    /**
     *
     * @type {string}
     * @memberof PaginatedEvalRunListList
     */
    next?: string | null;
    /**
     *
     * @type {string}
     * @memberof PaginatedEvalRunListList
     */
    previous?: string | null;
    /**
     *
     * @type {Array<EvalRunList>}
     * @memberof PaginatedEvalRunListList
     */
    results: Array<EvalRunList>;
}

/**
 * Check if a given object implements the PaginatedEvalRunListList interface.
 */
export function instanceOfPaginatedEvalRunListList(value: object): value is PaginatedEvalRunListList {
    if (!('count' in value) || value['count'] === undefined) return false;
    if (!('results' in value) || value['results'] === undefined) return false;
    return true;
}

export function PaginatedEvalRunListListFromJSON(json: any): PaginatedEvalRunListList {
    return PaginatedEvalRunListListFromJSONTyped(json, false);
}

export function PaginatedEvalRunListListFromJSONTyped(json: any, ignoreDiscriminator: boolean): PaginatedEvalRunListList {
    if (json == null) {
        return json;
    }
    return {

        'count': json['count'],
        'next': json['next'] == null ? undefined : json['next'],
        'previous': json['previous'] == null ? undefined : json['previous'],
        'results': ((json['results'] as Array<any>).map(EvalRunListFromJSON)),
    };
}

export function PaginatedEvalRunListListToJSON(json: any): PaginatedEvalRunListList {
    return PaginatedEvalRunListListToJSONTyped(json, false);
}

export function PaginatedEvalRunListListToJSONTyped(value?: PaginatedEvalRunListList | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {

        'count': value['count'],
        'next': value['next'],
        'previous': value['previous'],
        'results': ((value['results'] as Array<any>).map(EvalRunListToJSON)),
    };
}