// frontend/src/openapi/models/PatchedEvalSetMemberUpdateRequest.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Overbae API
 * Backend API for the Overclaw / Overmind platform.
 *
 * The version of the OpenAPI document: 2.0.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { RoleAbbEnum } from './RoleAbbEnum';
import {
    RoleAbbEnumFromJSON,
    RoleAbbEnumFromJSONTyped,
    RoleAbbEnumToJSON,
    RoleAbbEnumToJSONTyped,
} from './RoleAbbEnum';

/**
 *
 * @export
 * @interface PatchedEvalSetMemberUpdateRequest
 */
export interface PatchedEvalSetMemberUpdateRequest {
    /**
     *
     * @type {RoleAbbEnum}
     * @memberof PatchedEvalSetMemberUpdateRequest
     */
    role?: RoleAbbEnum;
    /**
     *
     * @type {boolean}
     * @memberof PatchedEvalSetMemberUpdateRequest
     */
    enabled?: boolean;
    /**
     *
     * @type {number}
     * @memberof PatchedEvalSetMemberUpdateRequest
     */
    order?: number;
}



/**
 * Check if a given object implements the PatchedEvalSetMemberUpdateRequest interface.
 */
export function instanceOfPatchedEvalSetMemberUpdateRequest(value: object): value is PatchedEvalSetMemberUpdateRequest {
    return true;
}

export function PatchedEvalSetMemberUpdateRequestFromJSON(json: any): PatchedEvalSetMemberUpdateRequest {
    return PatchedEvalSetMemberUpdateRequestFromJSONTyped(json, false);
}

export function PatchedEvalSetMemberUpdateRequestFromJSONTyped(json: any, ignoreDiscriminator: boolean): PatchedEvalSetMemberUpdateRequest {
    if (json == null) {
        return json;
    }
    return {

        'role': json['role'] == null ? undefined : RoleAbbEnumFromJSON(json['role']),
        'enabled': json['enabled'] == null ? undefined : json['enabled'],
        'order': json['order'] == null ? undefined : json['order'],
    };
}

export function PatchedEvalSetMemberUpdateRequestToJSON(json: any): PatchedEvalSetMemberUpdateRequest {
    return PatchedEvalSetMemberUpdateRequestToJSONTyped(json, false);
}

export function PatchedEvalSetMemberUpdateRequestToJSONTyped(value?: PatchedEvalSetMemberUpdateRequest | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {

        'role': RoleAbbEnumToJSON(value['role']),
        'enabled': value['enabled'],
        'order': value['order'],
    };
}